backup/nicayne_os_full_backup_2025_06_01/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fitz>=0.0.1.dev2",
    "flask>=3.1.1",
    "google-api-python-client>=2.170.0",
    "google-auth>=2.40.2",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "gspread>=6.2.1",
    "openai>=1.82.1",
    "pillow>=11.2.1",
    "pymupdf>=1.26.0",
    "pytesseract>=0.3.13",
    "qrcode[pil]>=8.2",
    "reportlab>=4.4.1",
    "requests>=2.32.3",
    "werkzeug>=3.1.3",
]
